# packages/api/src/generated/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/attachments/upload-url:
    post:
      operationId: AttachmentController_createUploadUrl
      parameters: []
      responses:
        '200':
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  upload_url:
                    type: string
                  s3_key:
                    type: string
                  expires_at:
                    type: number
        '401':
          description: Unauthorized - User not found
        '403':
          description: Forbidden - Access denied to household
      security:
        - session: []
      summary: Generate presigned URL for attachment upload
      tags: &ref_0
        - attachments
  /api/attachments/{id}/uploaded:
    post:
      operationId: AttachmentController_confirmUpload
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Upload confirmed successfully
        '401':
          description: Unauthorized - User not found
        '403':
          description: Forbidden - Access denied to household
        '404':
          description: Not Found - Attachment not found
      security:
        - session: []
      summary: Confirm successful attachment upload
      tags: *ref_0
  /api/attachments/{id}/download-url:
    get:
      operationId: AttachmentController_createDownloadUrl
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Presigned download URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  download_url:
                    type: string
                  expires_at:
                    type: number
        '401':
          description: Unauthorized - User not found
        '403':
          description: Forbidden - Access denied to household
        '404':
          description: Not Found - Attachment not found
      security:
        - session: []
      summary: Generate presigned URL for attachment download
      tags: *ref_0
  /api/attachments/{id}:
    delete:
      operationId: AttachmentController_deleteAttachment
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Attachment deleted successfully
        '401':
          description: Unauthorized - User not found
        '403':
          description: Forbidden - Access denied to household
        '404':
          description: Not Found - Attachment not found
      security:
        - session: []
      summary: Delete attachment and associated file
      tags: *ref_0
...
info:
  title: Pantry API
  description: API for collaborative pantry planning
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3001
    description: Local development server
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: pantry.session_token
      description: Session cookie from Better Auth
    powersync-jwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: PowerSync JWT token for data synchronization
...